Clinical informatician review and suggestions
I have added suggestions to text re the static population, the uptick in medication reviews and blood test based measures and just a note that one of the refsets for medication reviews is no longer active. 
It would be nice to quantify increase if possible, using the 'substantial sustained format' would that be straightforward to add? I think this was on the last report.